+++ ChangeLog.md
@@ -2,6 +2,7 @@
 
 ## current
 
+- implemented Data.Prime.isComposite
 - minor refactoring
 
 ## 0.1.0.1